+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Pause replay in session mode when offline (#4264)
 - Add replay quality option for Objective-C (#4267)
+- Expose span baggage API (#4207)
 
 ### Fixes